requirements.txt
PyDispatcher==2.0.3
pyserial==2.6
xively-python==0.1.0-rc1
gdata==2.0.18
